requirements.txt
akshare
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
tabulate>=0.9.0
loguru>=0.7.0
akquant>=0.1.0
# Web 后端
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
bcrypt>=4.0.0
